Macro ray_tracing::impl_vec_common

macro_rules! impl_vec_common {
    (named_struct@$fieldcount:literal, $Vec:ident($($param:tt),+), $($field:tt),+) => { ... };
    (tuple_struct@$fieldcount:literal, $Vec:ident($($param:tt),+), ($($field:tt),+)) => { ... };
    (common $fieldcount:literal, $Vec:ident($($param:tt),+), ($($field:tt),+)) => { ... };
}

Implements common traits and methods for a given vector.

1st arg is the representation (named struct or tuple struct) and the dimension of the given vector.

2nd arg is the identifier and parameters used to generate a new method; each parameter can be a single alphabet letter.

3rd arg is either the fields or the the indexes of the given vector, depending on its internal representation.

In order to implement dimension-specifc methods for your vector, see impl_vec_specific.

Example

// A tuple struct vector.
#[derive(Copy, Clone, Debug, Default, PartialEq)]
pub struct Vec3<T>(T, T, T);

impl_vec_common!(tuple_struct@3, Vec3(x, y, z), (0, 1, 2));
impl_vec_specific!(@3, Vec3);

// A named struct vector.
#[derive(Copy, Clone, Debug, Default, PartialEq)]
pub struct Vec4<T> {
    x: T,
    y: T,
    z: T,
    w: T,
}

impl_vec_common!(named_struct@4, Vec4(x, y, z, w), (x, y, z, w));
impl_vec_specific!(@4, Vec4);